+++ tests/Unit/Structs/Soa/v1/InitiateSignatureRequestTest.php
@@ -2,8 +2,11 @@
 
 namespace Coverzen\Components\YousignClient\Tests\Unit\Structs\Soa\v1;
 
+use Coverzen\Components\YousignClient\Enums\v1\DeliveryMode;
 use Coverzen\Components\YousignClient\Exceptions\Structs\v1\StructSaveException;
 use Coverzen\Components\YousignClient\Structs\Soa\v1\InitiateSignatureRequest;
+use Coverzen\Components\YousignClient\YousignClientServiceProvider;
+use Illuminate\Support\Facades\Config;
 
 /**
  * Class InitiateSignatureRequestTest.
@@ -12,6 +15,9 @@ use Coverzen\Components\YousignClient\Structs\Soa\v1\InitiateSignatureRequest;
  */
 final class InitiateSignatureRequestTest extends TestCase
 {
+    /** @var string */
+    private const FAKE_CUSTOM_EXPERIENCE_ID = 'fake-custom-experience-id';
+
     /**
      * @test
      *
@@ -42,19 +48,56 @@ final class InitiateSignatureRequestTest extends TestCase
 
     /**
      * @test
+     * @dataProvider customerExperienceProvider
+     *
+     * @param bool $customExperience
      *
      * @return void
      */
-    public function it_has_expected_properties(): void
+    public function it_has_expected_properties(bool $customExperience): void
     {
+        if ($customExperience) {
+            Config::set(YousignClientServiceProvider::CONFIG_KEY . '.custom_experience_id', self::FAKE_CUSTOM_EXPERIENCE_ID);
+        }
+
         /** @var InitiateSignatureRequest $initiateSignatureRequest */
         $initiateSignatureRequest = InitiateSignatureRequest::factory()
                                                             ->make();
-
+        $this->assertNotNull($initiateSignatureRequest->name);
         $this->assertIsString($initiateSignatureRequest->name);
-        $this->assertIsString($initiateSignatureRequest->delivery_mode);
+
+        $this->assertNotNull($initiateSignatureRequest->delivery_mode);
+        $this->assertInstanceOf(DeliveryMode::class, $initiateSignatureRequest->delivery_mode);
+
+        $this->assertNotNull($initiateSignatureRequest->ordered_signers);
         $this->assertIsBool($initiateSignatureRequest->ordered_signers);
+
+        $this->assertNotNull($initiateSignatureRequest->timezone);
         $this->assertIsString($initiateSignatureRequest->timezone);
+
+        $this->assertIsArray($initiateSignatureRequest->email_notification);
+
+        if ($customExperience) {
+            $this->assertNotNull($initiateSignatureRequest->custom_experience_id);
+        } else {
+            $this->assertNull($initiateSignatureRequest->custom_experience_id);
+        }
+    }
+
+    /**
+     * @test
+     *
+     * @return void
+     */
+    public function it_has_default_properties_values(): void
+    {
+        Config::set(YousignClientServiceProvider::CONFIG_KEY . '.custom_experience_id', self::FAKE_CUSTOM_EXPERIENCE_ID);
+
+        /** @var InitiateSignatureRequest $initiateSignatureRequest */
+        $initiateSignatureRequest = new InitiateSignatureRequest();
+
+        $this->assertSame(DeliveryMode::none(), $initiateSignatureRequest->delivery_mode);
+        $this->assertSame(Config::get(YousignClientServiceProvider::CONFIG_KEY . '.custom_experience_id'), $initiateSignatureRequest->custom_experience_id);
     }
 
     /**
@@ -82,5 +125,100 @@ final class InitiateSignatureRequestTest extends TestCase
         $this->expectException(StructSaveException::class);
 
         (new InitiateSignatureRequest())->save();
+    }
+
+    /**
+     * @test
+     * @dataProvider customerExperienceProvider
+     *
+     * @param bool $customExperience
+     *
+     * @return void
+     */
+    public function it_has_payload_accessor(bool $customExperience): void
+    {
+        if ($customExperience) {
+            Config::set(YousignClientServiceProvider::CONFIG_KEY . '.custom_experience_id', self::FAKE_CUSTOM_EXPERIENCE_ID);
+        }
+
+        /** @var InitiateSignatureRequest $initiateSignatureRequest */
+        $initiateSignatureRequest = InitiateSignatureRequest::factory()
+                                                            ->make();
+
+        $this->assertNotNull($initiateSignatureRequest->payload);
+        $this->assertIsArray($initiateSignatureRequest->payload);
+
+        $this->assertArrayNotHasKey('payload', $initiateSignatureRequest->payload);
+
+        $this->assertArrayHasKey('name', $initiateSignatureRequest->payload);
+        $this->assertArrayHasKey('delivery_mode', $initiateSignatureRequest->payload);
+        $this->assertArrayHasKey('ordered_signers', $initiateSignatureRequest->payload);
+        $this->assertArrayHasKey('timezone', $initiateSignatureRequest->payload);
+        $this->assertArrayHasKey('email_notification', $initiateSignatureRequest->payload);
+
+        if ($customExperience) {
+            $this->assertArrayHasKey('custom_experience_id', $initiateSignatureRequest->payload);
+        } else {
+            $this->assertArrayNotHasKey('custom_experience_id', $initiateSignatureRequest->payload);
+        }
+    }
+
+    /**
+     * Provides a set of.
+     *
+     * @return array<string,array<string,mixed>>
+     */
+    public static function customerExperienceProvider(): array
+    {
+        return [
+            'custom experience' => [
+                'customExperience' => true,
+            ],
+            'no custom experience' => [
+                'customExperience' => false,
+            ],
+        ];
+    }
+
+    /**
+     * @test
+     * @dataProvider customerExperienceProvider
+     *
+     * @param bool $customExperience
+     *
+     * @return void
+     */
+    public function it_removes_null_values_in_payload_accessor(bool $customExperience): void
+    {
+        if ($customExperience) {
+            Config::set(YousignClientServiceProvider::CONFIG_KEY . '.custom_experience_id', self::FAKE_CUSTOM_EXPERIENCE_ID);
+        }
+
+        /** @var InitiateSignatureRequest $initiateSignatureRequest */
+        $initiateSignatureRequest = InitiateSignatureRequest::factory()
+                                                            ->make(
+                                                                [
+                                                                    'name' => null,
+                                                                    'ordered_signers' => null,
+                                                                    'timezone' => null,
+                                                                    'email_notification' => null,
+                                                                ]
+                                                            );
+
+        $this->assertNotNull($initiateSignatureRequest->payload);
+        $this->assertIsArray($initiateSignatureRequest->payload);
+
+        $this->assertArrayNotHasKey('payload', $initiateSignatureRequest->payload);
+
+        $this->assertArrayNotHasKey('name', $initiateSignatureRequest->payload);
+        $this->assertArrayNotHasKey('ordered_signers', $initiateSignatureRequest->payload);
+        $this->assertArrayNotHasKey('timezone', $initiateSignatureRequest->payload);
+        $this->assertArrayNotHasKey('email_notification', $initiateSignatureRequest->payload);
+
+        if ($customExperience) {
+            $this->assertArrayHasKey('custom_experience_id', $initiateSignatureRequest->payload);
+        } else {
+            $this->assertArrayNotHasKey('custom_experience_id', $initiateSignatureRequest->payload);
+        }
     }
 }