+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+# [1.4.0](https://github.com/coverzen/yousign-client/compare/v1.3.1...v1.4.0) (2024-10-28)
+
+
+### Features
+
+* add custom experience id management ([1d5baea](https://github.com/coverzen/yousign-client/commit/1d5baea028e7e15e4f7e1e1a329af0552e758764))
+* add enum for delivery mode ([4c826d7](https://github.com/coverzen/yousign-client/commit/4c826d7716dd37445931a5322484181519b85667))
+* improve fake facade with inspection ([4d7fd8e](https://github.com/coverzen/yousign-client/commit/4d7fd8ebdf3519fce993be1acd8a7c395064719e))
+* improve InitiateSignatureRequest ([7631044](https://github.com/coverzen/yousign-client/commit/7631044c53469b50062630d9bd48c99bee586981))
+
+## [1.3.1](https://github.com/coverzen/yousign-client/compare/v1.3.0...v1.3.1) (2024-10-22)
+
+
+### Bug Fixes
+
+* improve yousign facade testing capabilities ([cedf67a](https://github.com/coverzen/yousign-client/commit/cedf67acfe34e8e28136efb4ec7bbe84ef6574d3))
+
+# [1.3.0](https://github.com/coverzen/yousign-client/compare/v1.2.0...v1.3.0) (2024-10-14)
+
+
+### Features
+
+* add facade to improve tests and package usage ([58eb30a](https://github.com/coverzen/yousign-client/commit/58eb30a3c02c70ec7ad3d7287d24ee72b68e821e))
+
 # [1.2.0](https://github.com/coverzen/yousign-client/compare/v1.1.0...v1.2.0) (2024-10-11)